requirements.txt
pandas==1.2.4
numpy==1.20.1
matplotlib==3.3.4
glob2==0.7
lmfit==1.0.3
tqdm==4.59.0
scikit-learn==0.24.1
python-dateutil==2.8.1
